Refactored. Added some comments. Renamed classes - base generic classes became the default Node and Heap classes. Certain implementation with Double type as a key class renamed with "DoubleKey" suffix.
